feat: create accordion page
